+++ tests/test_assertions.py
@@ -6,7 +6,6 @@ from contextlib import contextmanager
 import pytest
 
 import slash
-from slash import should
 from slash.exceptions import TestFailed as _TestFailed, ExpectedExceptionNotCaught
 
 
@@ -14,7 +13,6 @@ from slash.exceptions import TestFailed as _TestFailed, ExpectedExceptionNotCaug
     (1, 1),
     (1, 1.00000001),
 ])
-@pytest.mark.usefixtures('disable_vintage_deprecations')
 def test_assert_almost_equal_positive(pair):
     a, b = pair
     slash.assert_almost_equal(a, b)
@@ -26,7 +24,6 @@ def test_assert_almost_equal_positive(pair):
     (1, 1.1, 0.5),
     (1.0001, 1.00009, 0.00002),
 ])
-@pytest.mark.usefixtures('disable_vintage_deprecations')
 def test_assert_almost_equal_positive_with_delta(combination):
     a, b, delta = combination
     slash.assert_almost_equal(a, b, delta)
@@ -37,84 +34,13 @@ def test_assert_almost_equal_positive_with_delta(combination):
     (1, 1.1, 0.00001),
     (1.0001, 1.00009, 0.000001),
 ])
-@pytest.mark.usefixtures('disable_vintage_deprecations')
 def test_assert_almost_equal_negative_with_delta(combination):
     a, b, delta = combination
     with pytest.raises(AssertionError):
         slash.assert_almost_equal(a, b, delta)
 
 
-@pytest.mark.usefixtures('disable_vintage_deprecations')
-def test_assert_equals():
-    with checking(should.equal, should.not_equal):
-        good(1, 1)
-        bad(1, 2)
-
-
-@pytest.mark.usefixtures('disable_vintage_deprecations')
-def test_assert_empty():
-    with checking(should.be_empty, should.not_be_empty):
-        good([])
-        good({})
-        good(set())
-        bad([1, 2, 3])
-        bad({1: 2})
-        bad(set([1, 2, 3]))
-
-
-@pytest.mark.usefixtures('disable_vintage_deprecations')
-def test_assert_isinstance():
-    with checking(should.be_a, should.not_be_a):
-        good(1, int)
-        good("a", str)
-        good({}, dict)
-        bad(1, 1)
-        bad(1, str)
-        bad(None, str)
-
-
-@pytest.mark.usefixtures('disable_vintage_deprecations')
-def test_is_none():
-    with checking(should.be_none, should.not_be_none):
-        good(None)
-        bad("None")
-
-
-@pytest.mark.usefixtures('disable_vintage_deprecations')
-def test_is():
-    obj = object()
-    with checking(should.be, should.not_be):
-        good(obj, obj)
-        bad(obj, object())
-        bad({}, {})
-
-
-@pytest.mark.usefixtures('disable_vintage_deprecations')
-def test_truth():
-    with checking(should.be_true, should.be_false):
-        good(True)
-        good("hello")
-        bad(False)
-        bad("")
-        bad({})
-
-
-@pytest.mark.usefixtures('disable_vintage_deprecations')
-def test_in():
-    with checking(should.be_in, should.not_be_in):
-        good(1, [1, 2, 3])
-        good("e", "hello")
-        bad(1, [])
-        bad("e", "fdfd")
-    with checking(should.contain, should.not_contain):
-        good([1, 2, 3], 1)
-        good("hello", "e")
-        bad("fdfdfd", "e")
-
-
-
-@pytest.mark.usefixtures('disable_vintage_deprecations')
-@pytest.mark.parametrize('func', [should.raise_exception, slash.assert_raises])
+@pytest.mark.parametrize('func', [slash.assert_raises])
 def test_assert_raises(func):
     thrown = CustomException()
     with func(CustomException) as caught:
@@ -134,11 +60,10 @@ def test_assert_raises(func):
         assert " not raised" in str(e)
         assert e.expected_types == (CustomException,)
     else:
-        assert False, "should.raise_exception allowed success"
+        assert False, "assert_raises allowed success"
 
 
-@pytest.mark.usefixtures('disable_vintage_deprecations')
-@pytest.mark.parametrize('func', [should.raise_exception, slash.assert_raises])
+@pytest.mark.parametrize('func', [slash.assert_raises])
 def test_assert_raises_multiple_exceptions(func):
     class CustomException1(Exception):
         pass
@@ -162,20 +87,6 @@ def test_assert_raises_multiple_exceptions(func):
         with func(exception_types):
             raise value
     assert caught.value is value
-
-
-
-
-@pytest.mark.usefixtures('disable_vintage_deprecations')
-def test_raises_exception_multiple_classes():
-    possible_exception_types = (CustomException, OtherException)
-    for x in possible_exception_types:
-        with should.raise_exception(possible_exception_types):
-            raise x()
-
-    with pytest.raises(ExpectedExceptionNotCaught):
-        with should.raise_exception((CustomException,)):
-            pass
 
 # boilerplate